Refactor: split postifx handling in postifx_extension
